+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
-- No changes yet.
+- Log a warning in case an import statement does not point to a file in the module or in a direct
+  dependency.
 
 ## [v1.19.0] - 2023-05-17